Moved Native Methods to: Interop/Kernel32.cs New usage is: using OpenHardwareMonitor.Interop; Kernel32.
